backend/requirements.txt
asgiref==3.3.4
Django==3.2.4
django-webpack-loader==1.1.0
djangorestframework==3.12.4
pytz==2021.1
PyYAML==5.4.1
sqlparse==0.4.1
uritemplate==3.0.1
